+++ src/tree/NonTerminal.php
@@ -6,10 +6,36 @@ use VovanVE\parser\common\TreeNodeInterface;
 
 class NonTerminal extends BaseObject implements TreeNodeInterface
 {
-    /** @var string */
+    /**
+     * @var string
+     * @deprecated Don't use outside directly - use getter
+     */
     public $name;
-    /** @var TreeNodeInterface[] */
+    /** @var string|null */
+    private $tag;
+    /**
+     * @var TreeNodeInterface[]
+     * @deprecated Don't use outside directly - use getter
+     */
     public $children;
+
+    /**
+     * @param string $name
+     * @param TreeNodeInterface[] $children
+     * @param string|null $tag
+     * @since 1.3.0
+     */
+    public function __construct($name, $children, $tag = null)
+    {
+        $this->name = $name;
+        $this->children = $children;
+        if (null !== $tag) {
+            $tag = (string)$tag;
+            if ('' !== $tag) {
+                $this->tag = $tag;
+            }
+        }
+    }
 
     /**
      * @inheritdoc
@@ -17,6 +43,15 @@ class NonTerminal extends BaseObject implements TreeNodeInterface
     public function getNodeName()
     {
         return $this->name;
+    }
+
+    /**
+     * @inheritdoc
+     * @since 1.3.0
+     */
+    public function getNodeTag()
+    {
+        return $this->tag;
     }
 
     /**
@@ -43,8 +78,7 @@ class NonTerminal extends BaseObject implements TreeNodeInterface
      */
     public function areChildrenMatch($nodeNames)
     {
-        $count = $this->getChildrenCount();
-        if (count($nodeNames) !== $count) {
+        if (count($nodeNames) !== $this->getChildrenCount()) {
             return false;
         }
 
@@ -65,7 +99,13 @@ class NonTerminal extends BaseObject implements TreeNodeInterface
      */
     public function dumpAsString($indent = '', $last = true)
     {
-        $out = $indent . ' `- ' . $this->name . PHP_EOL;
+        $out = $indent . ' `- ' . $this->name;
+
+        if (null !== $this->tag) {
+            $out .= '(' . $this->tag . ')';
+        }
+
+        $out .= PHP_EOL;
         $sub_indent = $indent . ($last ? '    ' : ' |  ');
         $last_i = count($this->children) - 1;
         foreach ($this->children as $i => $child) {