+++ tests/unit/table/TableTest.php
@@ -164,7 +164,10 @@ _GRAMMAR
 
         $rule = $row->reduceRule;
         $this->assertInstanceOf(Rule::class, $rule);
-        $this->assertEquals(0, Rule::compare($rule, new Rule(new Symbol('S'), [new Symbol('E')], true)));
+        $this->assertEquals(0, Rule::compare(
+            $rule,
+            new Rule(new Symbol('S'), [new Symbol('E')], true)
+        ));
 
         return $terminals;
     }