+++ tests/unit/tree/NonTerminalTest.php
@@ -14,9 +14,7 @@ class NonTerminalTest extends BaseTestCase
     {
         $token = new Token('int', '42');
 
-        $node = new NonTerminal();
-        $node->name = 'V';
-        $node->children = [$token];
+        $node = new NonTerminal('V', [$token]);
 
         $this->assertEquals(1, $node->getChildrenCount());
         $this->assertCount(1, $node->getChildren());
@@ -56,6 +54,27 @@ DUMP
         return $node;
     }
 
+    public function testVTagInt()
+    {
+        $token = new Token('int', '37');
+
+        $node = new NonTerminal('V', [$token], 'num');
+
+        $this->assertEquals(1, $node->getChildrenCount());
+        $this->assertCount(1, $node->getChildren());
+
+        $this->assertEquals(<<<'DUMP'
+ `- V(num)
+     `- int <37>
+
+DUMP
+            ,
+            $node->dumpAsString()
+        );
+
+        return $node;
+    }
+
     /**
      * @param NonTerminal $v
      * @return NonTerminal
@@ -63,9 +82,7 @@ DUMP
      */
     public function testPVInt($v)
     {
-        $node = new NonTerminal();
-        $node->name = 'P';
-        $node->children = [$v];
+        $node = new NonTerminal('P', [$v]);
 
         $this->assertEquals(
             <<<'DUMP'
@@ -93,15 +110,13 @@ DUMP
      * @param NonTerminal $p
      * @param NonTerminal $v
      * @depends testPVInt
-     * @depends testVInt
+     * @depends testVTagInt
      */
     public function testEVMulInt($p, $v)
     {
         $mul = new Token('mul', '*');
 
-        $node = new NonTerminal();
-        $node->name = 'E';
-        $node->children = [$p, $mul, $v];
+        $node = new NonTerminal('E', [$p, $mul, $v]);
 
         $this->assertEquals(3, $node->getChildrenCount());
         $this->assertCount(3, $node->getChildren());
@@ -113,8 +128,8 @@ DUMP
 .... |   |   `- V
 .... |   |       `- int <42>
 .... |   `- mul <*>
-.... |   `- V
-.... |       `- int <42>
+.... |   `- V(num)
+.... |       `- int <37>
 
 DUMP